libcheck: Sanitize use of the HAVE_FORK define
The value of the HAVE_FORK macro was being checked inconsistently throughout
libcheck. In some places defined(HAVE_FORK) was assumed to mean that fork was
available and in other places HAVE_FORK==1 was required with HAVE_FORK==0
meaning fork was not available. Obviously this is wrong since if HAVE_FORK==0,
defined(HAVE_FORK) will also be true.